ENG-13: remove sign up folder, and remove the sign up provider
