+++ app/(auth)/_layout.tsx
@@ -9,7 +9,8 @@ export default function AuthLayout() {
       }}
     >
       <Stack.Screen name="index" />
-      <Stack.Screen name="(signUp)" />
+      <Stack.Screen name="signUp" />
+      <Stack.Screen name="createProfile" />
       <Stack.Screen name="forgotPassword" />
     </Stack>
   );
